# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Coordinates {
  lat: number;
  lng: number;
}

export interface Destination {
  id: number;
  name: string;
  slug: string;
  region: string;
  category: string;
  description: string;
  shortDescription: string;
  imageUrl: string;
  highlights: string[];
  bestTimeToVisit: string;
  difficulty?: string;
  duration?: string;
  featured: boolean;
  rating: number;
  reviewCount: number;
  coordinates?: Coordinates;
}

export interface Experience {
  id: number;
  title: string;
  category: string;
  description: string;
  imageUrl: string;
  duration: string;
  difficulty?: string;
  price: string;
  highlights: string[];
  included?: string[];
  destinationId?: number;
}

export interface Region {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  destinationCount: number;
  highlights: string[];
  climate: string;
}

export interface CreateInquiry {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  destination?: string;
  travelDate?: string;
  /** @minimum 1 */
  groupSize?: number;
  /** @minLength 10 */
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string;
  destination?: string;
  travelDate?: string;
  groupSize?: number;
  message: string;
  createdAt: string;
}

export type GetDestinationsParams = {
  region?: string;
  category?: string;
  featured?: boolean;
};

export type GetExperiencesParams = {
  category?: string;
};