# lib/api-zod/src/generated/types/createInquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInquiry {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  destination?: string;
  travelDate?: string;
  /** @minimum 1 */
  groupSize?: number;
  /** @minLength 10 */
  message: string;
}