# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinates } from "./coordinates";

export interface Destination {
  id: number;
  name: string;
  slug: string;
  region: string;
  category: string;
  description: string;
  shortDescription: string;
  imageUrl: string;
  highlights: string[];
  bestTimeToVisit: string;
  difficulty?: string;
  duration?: string;
  featured: boolean;
  rating: number;
  reviewCount: number;
  coordinates?: Coordinates;
}