# lib/api-zod/src/generated/types/experience.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Experience {
  id: number;
  title: string;
  category: string;
  description: string;
  imageUrl: string;
  duration: string;
  difficulty?: string;
  price: string;
  highlights: string[];
  included?: string[];
  destinationId?: number;
}