# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string;
  destination?: string;
  travelDate?: string;
  groupSize?: number;
  message: string;
  createdAt: string;
}