# lib/api-zod/src/generated/types/region.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algeria Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Region {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  destinationCount: number;
  highlights: string[];
  climate: string;
}